requirements.txt
flask
requests
python-dotenv
feedparser
beautifulsoup4